Support defaulted parameters to other commands.
Switched from enumerating ``self.distribution.commands`` to using
``self.distribution.get_command_list``.  The former is the list of
command names that will be run and the latter is the list of all
known commands.  This will make the code work when you do not
include something like ``sdist`` in the same command list.